pyproject.toml
[project]
name = "asynchronous-agents"
version = "0.1.0"
description = ""
authors = [
    {name = "Philipp Comans",email = "philipp.comans@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pika (>=1.3.2,<2.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "langgraph (>=0.3.24,<0.4.0)",
    "langchain-openai (>=0.3.12,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
